Issues #19 (apache_get_modules doesn't work always) and #20 (resolved: MysqlConnector must use pconnect instead of connect as default) resolved. Improved FileManager Class documentation.

+++ index.php
@@ -26,14 +26,15 @@ if(!class_exists('core')){
 	class core{
 
 		static function startup(){
-			
-			/*
-			 * Issue 0000002: http://alpha.diegolamonica.info/issues/view.php?id=2
-			 */
-			$ext = apache_get_modules();
-			if(!array_search('mod_rewrite', $ext)){
-				echo('Apache Module: <strong>mod_rewrite</strong> not loaded.');
-				exit();
+			if(function_exists('apache_get_modules')){
+				/*
+				 * Issue 0000002: http://alpha.diegolamonica.info/issues/view.php?id=2
+				 */
+				$ext = apache_get_modules();
+				if(!array_search('mod_rewrite', $ext)){
+					echo('Apache Module: <strong>mod_rewrite</strong> not loaded.');
+					exit();
+				}
 			}
 			/*
 			 * Security issue, blocking to override the __fn and the __url $_GET variables
